crypto: qat - lce: add mailbox communication with CPF
Introduce the definition of mailbox messages exchanged between the
LCE PF driver and its counterpart, the CPF driver, which operates
within the IPU internal compute complex.

The LCE mailbox consists of a pair of 32-bit registers with
interrupt-to-peer capability. BIT(0) generates an interrupt, the
remaining bits are software defined. PF<->CPF and VF<->CPF use the
same encoding with separate message-type spaces.

Signed-off-by: Tanvi Desai <tanvi.desai@intel.com>