misc: lce: Add definition of mailbox messages for RPC
Introducing the definition of mailbox messages exchanged
between the LCE PF driver and its counterpart, the CPF driver,
which operates within the IPU internal compute complex.

The exchange of mailbox messages occurs through a straightforward
mechanism utilizing a pair of 32-bit registers accessible from two
sides, with the capability of generating interrupts to the other side.
Requestor writes a message, while the other side reads it, and optionally
responds with another message. The arrival of new message or response
is notified through interrupts.

The same scheme is applicable for message exchange between VF driver
on Host and LCE CPF driver as well.

In addition to the aforementioned registers, two additional 32-bit message
registers are available to the PF to accommodate larger payloads exchange
with LCE CPF. Use of this additional registers depends on specific message
opcode in the original request or reply.

The RPC protocol is implemented in next patch.

Signed-off-by: Tanvi Desai <tanvi.desai@intel.com>